move mini reports/votes to common components
for watchlist usage
