+++ src/components/common/VotingsMiniSection.tsx
@@ -1,24 +1,27 @@
 import { Stack, Text, Card, Group, Badge, Tooltip } from "@mantine/core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { IVoting } from "../../../../interfaces/Voting";
-import UserLink from "../../common/UserLink";
-import DateBadge from "../../common/badges/DateBadge";
-import { TruncatedText } from "../../common/TruncatedText";
+import { IVoting } from "../../../interfaces/Voting";
+import UserLink from "./UserLink";
+import DateBadge from "./badges/DateBadge";
+import { TruncatedText } from "./TruncatedText";
 
 interface IProps {
     votings: IVoting[];
+    hideTooltip?: boolean;
 }
 
-export default function TournamentVotings({ votings }: IProps) {
+export default function VotingsMiniSection({ votings, hideTooltip = false }: IProps) {
     return (
         <Stack gap={5}>
             <Group gap="xs" align="center">
                 <Text size="sm" fw={500} className="header-border-left">
                     Votes
                 </Text>
-                <Tooltip multiline w={232} label="Only shows votes matching the tournament's name or forum URL">
-                    <FontAwesomeIcon icon="exclamation-circle" size="sm" style={{ opacity: 0.6 }} />
-                </Tooltip>
+                {!hideTooltip && (
+                    <Tooltip multiline w={232} label="Only shows votes matching the tournament's name or forum URL">
+                        <FontAwesomeIcon icon="exclamation-circle" size="sm" style={{ opacity: 0.6 }} />
+                    </Tooltip>
+                )}
                 <Badge color={votings.length > 0 ? "blue" : "gray"} variant="light" size="sm">
                     {votings.length}
                 </Badge>
